+++ CHANGELOG.md
@@ -36,6 +36,7 @@ RELEASING:
 ## [Unreleased]
 ### Added
 - API endpoint "centrality" to calculate [betweenness centrality](https://en.wikipedia.org/wiki/Betweenness_centrality) values for nodes inside a given bounding box. Centrality is calculated using Brandes' algorithm. 
+- Support for turn restrictions with core-based routing algorithms
 ### Changed
 - Use Querygraph and virtual nodes for isochrone calculation in the same manner as in routing
 - Remove Isochrones v1 api tests